Ex003

+++ Exercícios/ex002/src/ex002.java
@@ -1,16 +1,16 @@
 import java.util.Scanner;
 
 public class ex002 {
-    public static void main(String[] args){
+    public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         
-        float x, y, z;
+        int x, y, z;
 
-        x = sc.nextFloat();
-        y = sc.nextFloat();
+        x = sc.nextInt();
+        y = sc.nextInt();
         z = x + y;
 
-        System.out.printf("A soma dos valor %.1f e %.1f resulta em %.1f", x, y, z);
+        System.out.printf("A soma dos valor %d e %d resulta em %d", x, y, z);
         
         sc.close();
     }
